pkg: Use cilium/ebpf code to read overwritable perf ring buffer.
Before this commit, we used specifically written code to read overwritable perf
ring buffers.
This code was upstreamed to cilium/ebpf and respects the already existing API
based on Read()/ReadInto() [1].

So, this commit modifies traceloop tracer code to use the cilium/ebpf API.

[1] https://github.com/cilium/ebpf/commit/5ea5368f9823c13e71146a6c2cca84ad60a5e26a
